requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile ./requirements/base.in
#
appdirs==1.4.4
    # via black
asgiref==3.5.0
    # via django
astroid==2.9.3
    # via
    #   pylint
    #   pylint-celery
    #   pylint-flask
    #   requirements-detector
backports-zoneinfo==0.2.1
    # via django
black==21.7b0
    # via blue
blue==0.7.0
    # via -r ./requirements/base.in
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.0.4
    # via
    #   black
    #   pip-compile-multi
    #   pip-tools
coreapi==2.3.3
    # via
    #   django-rest-swagger
    #   openapi-codec
coreschema==0.0.4
    # via coreapi
django==4.0.3
    # via
    #   django-filter
    #   djangorestframework
django-filter==21.1
    # via -r ./requirements/base.in
django-rest-swagger==2.2.0
    # via -r ./requirements/base.in
djangorestframework==3.13.1
    # via
    #   -r ./requirements/base.in
    #   django-rest-swagger
dodgy==0.2.1
    # via prospector
flake8==3.8.4
    # via
    #   blue
    #   flake8-polyfill
flake8-polyfill==1.0.2
    # via pep8-naming
idna==3.3
    # via requests
importlib-metadata==4.11.3
    # via markdown
isort==5.10.1
    # via pylint
itypes==1.2.0
    # via coreapi
jinja2==3.0.3
    # via coreschema
lazy-object-proxy==1.7.1
    # via astroid
markdown==3.3.6
    # via -r ./requirements/base.in
markupsafe==2.1.0
    # via jinja2
mccabe==0.6.1
    # via
    #   flake8
    #   prospector
    #   pylint
mypy-extensions==0.4.3
    # via black
openapi-codec==1.3.2
    # via django-rest-swagger
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pep8-naming==0.10.0
    # via prospector
pip-compile-multi==2.4.3
    # via -r ./requirements/base.in
pip-tools==6.5.1
    # via pip-compile-multi
platformdirs==2.5.1
    # via pylint
prospector==1.6.0
    # via -r ./requirements/base.in
pycodestyle==2.6.0
    # via
    #   flake8
    #   prospector
pydocstyle==6.1.1
    # via prospector
pyflakes==2.2.0
    # via
    #   flake8
    #   prospector
pylint==2.12.2
    # via
    #   prospector
    #   pylint-celery
    #   pylint-django
    #   pylint-flask
    #   pylint-plugin-utils
pylint-celery==0.3
    # via prospector
pylint-django==2.5.2
    # via prospector
pylint-flask==0.6
    # via prospector
pylint-plugin-utils==0.7
    # via
    #   prospector
    #   pylint-celery
    #   pylint-django
    #   pylint-flask
python-decouple==3.6
    # via -r ./requirements/base.in
pytz==2021.3
    # via djangorestframework
pyyaml==6.0
    # via prospector
regex==2022.3.2
    # via black
requests==2.27.1
    # via coreapi
requirements-detector==0.7
    # via prospector
setoptconf-tmp==0.3.1
    # via prospector
simplejson==3.17.6
    # via django-rest-swagger
snowballstemmer==2.2.0
    # via pydocstyle
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via
    #   prospector
    #   pylint
tomli==1.2.3
    # via
    #   black
    #   pep517
toposort==1.7
    # via pip-compile-multi
typing-extensions==4.1.1
    # via
    #   astroid
    #   pylint
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.8
    # via requests
wheel==0.37.1
    # via pip-tools
wrapt==1.13.3
    # via astroid
zipp==3.7.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
